user queries from src.queries
