+++ src/service/queries/__init__.py
@@ -1,6 +1,12 @@
 from typing import List
 
 from .role_queries import get_role_name_by_id
-from .user_queries import query_user_by_id
+from .user_queries import (get_all_users_by_role_id, insert_user,
+                           query_user_by_id)
 
-__all__: List[str] = ["get_role_name_by_id", "query_user_by_id"]
+__all__: List[str] = [
+    "get_role_name_by_id",
+    "query_user_by_id",
+    "get_all_users_by_role_id",
+    "insert_user",
+]
